% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembly.R
\name{add_processor}
\alias{add_processor}
\title{Add a Processor Layer to an Assembly Instruction Set}
\usage{
add_processor(x)
}
\arguments{
\item{x}{A vector of instructions}
}
\value{
A list of vectors of length two. The first item is \code{"proc"} and the
second is the original instruction in x.
}
\description{
Add a Processor Layer to an Assembly Instruction Set
}
