% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{create_assembly}
\alias{create_assembly}
\title{Create an AoC Assembly Computer}
\usage{
create_assembly(registers, functions, increment = 1, regex = "[, \\\\+]+")
}
\arguments{
\item{registers}{A list of registers}

\item{functions}{A list of functions. Registers can be referred to with
\code{self[[register_name]]}. They must have two arguments, \code{x} and \code{y}, even if
only one is used. The index is iterated with \code{private$.inc(j)} where \code{j} is
the number of places to "jump." Example: \verb{\\(x, y) private$.inc(self[[x]])}
will increment \code{self$index} by \code{self[[x]]}.}

\item{increment}{(Default: \code{1}) The number of places to increment by default.}

\item{regex}{(Default: \code{"[, \\\\+]+"}) The pattern to use for
\link[stringr:str_split]{stringr::str_split}}
}
\value{
A new R6 object with \code{registers} and \code{functions} as the fields and
methods, respectively. Along with the functions detailed in the details
section.
}
\description{
Create an AoC Assembly Computer
}
\details{
\code{try_numeric(x)}
:  Takes input and tries to cast it as numeric. If that fails, it returns the
original object.

\code{val_or_index(x)}
:  If \code{x} is numeric, return \code{x}, otherwise return \code{self[[x]]}.

\code{call(f, x, y)}
:  Calls one of \code{functions} using \code{x} and \code{y} as arguments.
There are two ways the function can be used:
\itemize{
\item \code{f} is not length one and \code{x} and \code{y} are null, then \code{f[1](f[2], f[3])}
\item \code{f} is length one, and \code{x} and \code{y} are not null, then \code{f(x, y)}
}

\code{run(x, target, until, pattern)}
:  Tries to run the instructions provided by \code{x}. If \code{x} is a character
vector, it will try to split the instructions into a segments for
\code{call(f, x, y)}.
\itemize{
\item \code{x}: A character vector of instructions as \code{c("f x y")} or a list of
character vectors in the format \code{list(c("f", "x", "y"))}
\item \code{target}: (Default: \code{NULL}) The \code{register} to print as the final result.
If \code{NULL}, \code{self} will be printed.
\item \code{pattern}: See the \code{regex} argument for more details.
}

\code{.inc(increment)}
:  A private function used to increment \code{self$index} by \code{increment}.
}
