% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{create_basic}
\alias{create_basic}
\title{Create a BASIC Computer}
\usage{
create_basic(registers, functions, increment = 1, regex = "[, \\\\+]+")
}
\arguments{
\item{registers}{A list of registers}

\item{functions}{A list of functions. Registers can be referred to with
\code{self[[register_name]]}. The index is iterated with \code{private$.inc(j)} where
\code{j} is the number of places to "jump."}
}
\value{
A new R6 object
}
\description{
Create a BASIC Computer
}
