% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{cast_matrix}
\alias{cast_matrix}
\title{Cast a matrix to the desired type}
\usage{
cast_matrix(x, fn)
}
\arguments{
\item{x}{A matrix}

\item{fn}{A bare function such as \link{as.character} or \link{as.integer}}
}
\value{
A matrix of the desired type
}
\description{
Cast a matrix to the desired type
}
