% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords_to_matrix.R
\name{coords_to_matrix}
\alias{coords_to_matrix}
\title{Convert a set of coords to a matrix}
\usage{
coords_to_matrix(
  data,
  row = data$row,
  col = data$col,
  maxRow = NULL,
  maxCol = NULL
)
}
\arguments{
\item{data}{A vector of data}

\item{row}{A vector of row indices}

\item{col}{A vector of column indices}

\item{maxRow}{The row count for the resulting matrix}

\item{maxCol}{The column count for the resulting matrix}
}
\value{
A matrix of dimensions row,col with data
}
\description{
Convert a set of coords to a matrix
}
