% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{.replace_na}
\alias{.replace_na}
\title{Replace NA}
\usage{
.replace_na(x)
}
\arguments{
\item{x}{A vector or matrix}
}
\value{
\code{x} but \code{NA} values have been replaced with \code{FALSE}
}
\description{
Replace NA
}
\keyword{internal}
