% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_directions.R
\name{expand_directions}
\alias{expand_directions}
\title{expand_directions}
\usage{
expand_directions(.data, preserve_data = FALSE)
}
\arguments{
\item{.data}{A data.frame or vector depending on \code{preserve_data}}

\item{preserve_data}{A boolean to indicate if a data.frame (TRUE) or vector
(FALSE) should be returned.}
}
\value{
A vector of directions
}
\description{
expand_directions
}
