% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash.R
\name{flash}
\alias{flash}
\title{Perform a Function on Each Cell in a Matrix}
\usage{
flash(m, .fn, i, cores = future::availableCores())
}
\arguments{
\item{m}{A matrix}

\item{.fn}{The function to apply}

\item{i}{The number of times to apply the function to the whole matrix}

\item{cores}{The amount of cores to use when applying the function to each
cell}
}
\value{
A matrix of the same dimensions as \code{m}
}
\description{
This performs the function without changing the matrix, then
applies all changes at once.
}
