% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_adjacent.R
\name{get_adjacent_coords}
\alias{get_adjacent_coords}
\title{Get coordinates adjacent to a point}
\usage{
get_adjacent_coords(
  yy,
  xx,
  nrow = NULL,
  ncol = NULL,
  data = NULL,
  yCount = 1,
  xCount = 1
)
}
\arguments{
\item{yy}{The target y coordinate}

\item{xx}{The target x coordinate}

\item{nrow}{The maximum row value. Not needed if \code{data} is provided.}

\item{ncol}{The maximum column value. Not needed if \code{data} is provided.}

\item{data}{A matrix used to determine \code{nrow} and \code{ncol} if provided.}

\item{yCount}{The number of rows to include on either side of the target}

\item{xCount}{The number of columns to include on either side of the target}
}
\value{
A matrix of row and col coordinates
}
\description{
Get coordinates adjacent to a point
}
\details{
Does NOT include the target point. If a wall is clipped, the
coordinates will not include the clipped wall.
}
