% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_adjacent.R
\name{get_adjacent_values}
\alias{get_adjacent_values}
\title{Get values adjacent to the target}
\usage{
get_adjacent_values(yy, xx, data, yCount = 1, xCount = 1)
}
\arguments{
\item{yy}{The target y coordinate}

\item{xx}{The target x coordinate}

\item{data}{A matrix}

\item{yCount}{The number of rows to include on either side of the target}

\item{xCount}{The number of columns to include on either side of the target}
}
\value{
A matrix of the dimensions specified
}
\description{
Get values adjacent to the target
}
\details{
Does include the target point. If a wall is clipped, the dimensions
will not include the clipped wall.
}
