% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_diagonal.R
\name{get_diagonal}
\alias{get_diagonal}
\title{Get any matrix diagonal}
\usage{
get_diagonal(m, x, dim = c("row", "col"), dir = c("up", "down"))
}
\arguments{
\item{m}{A matrix}

\item{x}{A row or column index corresponding to the chosen \code{dim}}

\item{dim}{\code{"row"} or \code{"col"} to indicate where \code{x} should be applied}

\item{dir}{Whether or not the diagonal goes \code{"up"} or \code{"down"}}
}
\value{
A vector of the diagonal
}
\description{
Get any matrix diagonal
}
