% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_diagonal.R
\name{get_diagonal2}
\alias{get_diagonal2}
\title{Get a list of matrix diagonals}
\usage{
get_diagonal2(m, corner = c("tl", "tr", "bl", "br"))
}
\arguments{
\item{m}{A square matrix}

\item{corner}{A coner where the diagonal should start from}
}
\value{
A list of vectors
}
\description{
Get a list of matrix diagonals
}
\details{
The returned vectors all start on the top or bottom of the original
matrix. If opposing corners (eg top left and bottom right) are chosen, then
the longest vector from the bottom corner will be dropped because it is a
duplicate of the longest from the top.
}
