% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{index}
\alias{index}
\title{Get the index of an object using modulo}
\usage{
index(x, n)
}
\arguments{
\item{x}{A vector or list}

\item{n}{A numeric vector of indices}
}
\value{
A numeric index
}
\description{
Get the index of an object using modulo
}
