% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_multiple.R
\name{lag_multiple}
\alias{lag_multiple}
\title{Compute a lagged version of the input data with multiple lags at once}
\usage{
lag_multiple(x, k = 1, name = NULL)
}
\arguments{
\item{x}{A vector, single column matrix, or univariate time series. Can also
be a multi-column matrix if \code{k} is length 1.}

\item{k}{(Optional) An integer vector containing a number of lags. Defaults
to 1}

\item{name}{(Optional) A name to be used in the lagged data.frame. Defaults
to the name of the variable passed to \code{x}. If that is not possible, \code{name}
will default to "X".}
}
\value{
Returns a data.frame of the lagged variable. The number of rows is
the same as the length of the input vector. The number of columns is the
number of lags to be used. Each column retains the name of the original
variable and includes the number of lags used for that column. If \code{x} is a
multi-column matrix, returns a matrix of the same number of columns with no
names.
}
\description{
Compute a lagged version of the input data with multiple lags at once
}
\examples{
# Creating dummy data
x <- rnorm(10)
# for lags 1-5
lag_multiple(x, 1:5)

# 1 lag with a matrix of dummy data
lag_multiple(matrix(1:100, 10, 10), 1)
}
