% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop.R
\name{loop}
\alias{loop}
\title{Offset a Vector in a Looping Fashion}
\usage{
loop(x, n, margin = c("row", "col"), index, index1 = FALSE)
}
\arguments{
\item{x}{A vector or a matrix}

\item{n}{The number of places to offset}

\item{margin}{"row" or "col" of a matrix}

\item{index}{The row or column number to change}

\item{index1}{Whether or not the index starts at 0 (\code{FALSE}) or 1 (\code{TRUE})}
}
\value{
\code{x}, but with the appropriate shifting
}
\description{
Offset a Vector in a Looping Fashion
}
