% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_day.R
\name{make_day}
\alias{make_day}
\title{Make a File for a New Day}
\usage{
make_day(
  year = lubridate::year(Sys.Date()),
  day = lubridate::mday(Sys.Date() + 1)
)
}
\arguments{
\item{year}{The year}

\item{day}{The day}
}
\value{
Nothing
}
\description{
Make a File for a New Day
}
