% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{matrix}
\alias{matrix}
\alias{matrix_intersect}
\alias{matrix_anti_join}
\alias{matrix_join}
\title{Matrix intersections}
\usage{
matrix_intersect(x, y, false = "", fn = function(x, y) x == y)

matrix_anti_join(x, y, false = "", fn = function(x, y) x == y)

matrix_join(x, y, false = "", fn = function(z) z != false)
}
\arguments{
\item{x}{A matrix}

\item{y}{A matrix}

\item{false}{The value to use when \code{fn} evaluates to \code{FALSE}}

\item{fn}{A function to compare the matrices}
}
\description{
Matrix intersections
}
\examples{
x <- matrix(sample(c('', '@'), 10, replace = TRUE), 2) |> print()
y <- matrix(sample(c('', '@'), 10, replace = TRUE), 2) |> print()

matrix_intersect(x, y)
matrix_join(x, y)
matrix_anti_join(x, y)
}
