% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_as.R
\name{matrix_as}
\alias{matrix_as}
\alias{as.integer.matrix}
\alias{as.character.matrix}
\alias{as.numeric.matrix}
\alias{as.logical.matrix}
\title{Convert a matrix to the given type}
\usage{
\method{as.integer}{matrix}(x)

\method{as.character}{matrix}(x)

\method{as.numeric}{matrix}(x)

\method{as.logical}{matrix}(x)
}
\arguments{
\item{x}{A matrix}
}
\value{
A matrix in the desired type
}
\description{
Convert a matrix to the given type
}
