% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_to_coords.R
\name{matrix_to_coords}
\alias{matrix_to_coords}
\title{Convert a matrix to a matrix of coordinates and values}
\usage{
matrix_to_coords(mat)
}
\arguments{
\item{mat}{A matrix to be converted to coordinate pair and data columns}
}
\value{
A data.frame with row, col, and data columns
}
\description{
Convert a matrix to a matrix of coordinates and values
}
