% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mistlecode-package.R
\docType{package}
\name{mistlecode-package}
\alias{mistlecode}
\alias{mistlecode-package}
\title{mistlecode: My Advent of Code helper package}
\description{
Provides a set of functions to tackle common Advent of Code problem styles. I can't promise they will help for future puzzles, but they may help for past ones.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://mistlecode.guslipkin.me/}
  \item \url{https://github.com/guslipkin/mistlecode}
  \item Report bugs at \url{https://github.com/guslipkin/mistlecode/issues}
}

}
\author{
\strong{Maintainer}: Gus Lipkin \email{gus.lipkin@gmail.com}

}
\keyword{internal}
