% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_grid.R
\name{parse_grid}
\alias{parse_grid}
\title{Parse a grid input}
\usage{
parse_grid(vec, width = NULL)
}
\arguments{
\item{vec}{A vector of lines of text created using readLines}

\item{width}{The width of each chunk. If not supplied, it is guessed (poorly...)}
}
\value{
A list of vectors representing the grid
}
\description{
Parse a grid input
}
