% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_list}
\alias{read_list}
\title{Read a list of vectors}
\usage{
read_list(path, split = " ", cast = identity)
}
\arguments{
\item{path}{A file path}

\item{split}{A character to split the rows by}

\item{cast}{A bare function to cast with}
}
\value{
A list of vectors of type \code{cast}
}
\description{
Read a list of vectors
}
