% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strg_extract_all.R
\name{strg_extract_all}
\alias{strg_extract_all}
\title{Actually Extract All Patterns}
\usage{
strg_extract_all(str, pattern)
}
\arguments{
\item{str}{A string or vector of strings}

\item{pattern}{A regex pattern. If you want to extract multiple terms, use
the regex or operator. (For 'one' or 'two': \code{one|two})}
}
\value{
A list of matches the same length as \code{str}. Each list item will
contain all matches.
}
\description{
Actually Extract All Patterns
}
\examples{
c(
  'oneightwo',
  '1two3four'
) |>
  strg_extract_all('one|two|three|four|five|six|seven|eight|nine|[1-9]')
}
