% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembly_utils.R
\name{add_processor}
\alias{add_processor}
\alias{get_premade}
\title{Add a Processor Layer to an Assembly Instruction Set}
\usage{
add_processor(x)

get_premade(
  x = c("jump", "set", "add", "subtract", "multiply", "divide", "modulo")
)
}
\arguments{
\item{x}{A vector of functions. The options are \code{c("set", "add", "subtract", "multiply", "divide", "modulo")}. A named vector can be supplied where the
functions returned will match the values of the supplied vector with the
names of the supplied vector.}
}
\value{
A list of vectors of length two. The first item is \code{"proc"} and the
second is the original instruction in x.

A list of functions.
}
\description{
Add a Processor Layer to an Assembly Instruction Set

Easily Get Premade Assembly Instructions
}
\examples{
\dontrun{
get_premade(c("add", "prod" = "multiply"))
}
}
