% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/follow_directions.R
\name{follow_directions}
\alias{follow_directions}
\title{follow_directions}
\usage{
follow_directions(.data, row = 0, col = 0, preserve_data = FALSE)
}
\arguments{
\item{.data}{A vector of \code{RLUD} directions. See \code{expand_directions} to
generate from a data.frame of direction and magnitude.}

\item{row}{(Default: \code{0}) The starting row value}

\item{col}{(Default: \code{0}) The starting column value}

\item{preserve_data}{A boolean to indicate if data besides \code{row} and \code{col}
should be returned.}
}
\value{
A data.frame of coordinate pairs with \code{row} and \code{col} values.
}
\description{
follow_directions
}
